Reorganize giant lib into separate node definitions.
