+++ html-node-core/Cargo.toml
@@ -1,28 +1,30 @@
 [package]
 name = "html-node-core"
-authors.workspace = true
-categories.workspace = true
+
+authors.workspace     = true
+categories.workspace  = true
 description.workspace = true
-edition.workspace = true
-homepage.workspace = true
-keywords.workspace = true
-license.workspace = true
-readme.workspace = true
-repository.workspace = true
-version.workspace = true
+edition.workspace     = true
+homepage.workspace    = true
+keywords.workspace    = true
+license.workspace     = true
+readme.workspace      = true
+repository.workspace  = true
+version.workspace     = true
 
     [package.metadata.docs.rs]
     all-features = true
     rustdoc-args = ["--cfg", "docsrs"]
 
 [dependencies]
-axum = { version = "0.6", optional = true, default-features = false }
+axum  = { version = "0.6", optional = true, default-features = false }
 serde = { version = "1.0", optional = true, features = ["derive"] }
 
 html-escape = "0.2"
-paste = "1.0.14"
+paste       = "1.0.14"
 
 [features]
-axum = ["dep:axum"]
-typed = []
-serde = ["dep:serde"]
+axum   = ["dep:axum"]
+pretty = []
+serde  = ["dep:serde"]
+typed  = []
