+++ html-node/Cargo.toml
@@ -1,40 +1,41 @@
 [package]
 name = "html-node"
-authors.workspace = true
-categories.workspace = true
-description.workspace = true
-edition.workspace = true
-homepage.workspace = true
-keywords.workspace = true
-license.workspace = true
-readme.workspace = true
-repository.workspace = true
-version.workspace = true
+
 documentation = "https://docs.rs/html-node"
+
+authors.workspace     = true
+categories.workspace  = true
+description.workspace = true
+edition.workspace     = true
+homepage.workspace    = true
+keywords.workspace    = true
+license.workspace     = true
+readme.workspace      = true
+repository.workspace  = true
+version.workspace     = true
 
     [package.metadata.docs.rs]
     all-features = true
     rustdoc-args = ["--cfg", "docsrs"]
 
-
 [[example]]
-name = "axum"
+name              = "axum"
 required-features = ["axum"]
 
 [[example]]
-name = "typed_custom_attributes"
+name              = "typed_custom_attributes"
 required-features = ["typed"]
 
 [dependencies]
-html-node-core = { version = "0.2", path = "../html-node-core" }
+html-node-core  = { version = "0.2", path = "../html-node-core" }
 html-node-macro = { version = "0.2", path = "../html-node-macro" }
 
-
 [dev-dependencies]
-axum = "0.6"
+axum  = "0.6"
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
 
 [features]
-axum = ["html-node-core/axum"]
-serde = ["html-node-core/serde"]
-typed = ["html-node-core/typed", "html-node-macro/typed"]
+axum   = ["html-node-core/axum"]
+pretty = ["html-node-core/pretty"]
+serde  = ["html-node-core/serde"]
+typed  = ["html-node-core/typed", "html-node-macro/typed"]
